# Optimize decoding by using CBOR StreamDecoder