--- a/runtime/interpreter/encoding_test.go
+++ b/runtime/interpreter/encoding_test.go
@@ -24,7 +24,6 @@ import (
 	"math/big"
 	"testing"
 
-	"github.com/fxamacker/cbor/v2"
 	"github.com/stretchr/testify/require"
 
 	"github.com/onflow/cadence/runtime/common"
@@ -5269,20 +5268,14 @@ func TestDecodeCallback(t *testing.T) {
 	require.Equal(t,
 		[]decodeCallback{
 			{
-				value: []interface{}{
-					cbor.Tag{
-						Number:  cborTagInt8Value,
-						Content: uint64(42),
-					},
-				},
-				path: nil,
+				value: Int8Value(42),
+				path:  []string{"0"},
 			},
 			{
-				value: cbor.Tag{
-					Number:  cborTagInt8Value,
-					Content: uint64(42),
+				value: &ArrayValue{
+					Values: []Value{Int8Value(42)},
 				},
-				path: []string{"0"},
+				path: nil,
 			},
 		},
 		decodeCallbacks,