Closes #863

## Description

### Changes
 - [x] Optimize decoding of Value objects by using StreamDecoder and low-level API in fxamacker/cbor
 - [x] Modify tests

Types using StreamDecoder  include:  ArrayValue, CompositeValue, DictionaryValue, StorageReferenceValue, PathValue, CapabilityValue, common.AddressLocation, LinkValue, CompositeStaticType, InterfaceStaticType, ConstantSizedStaticType, ReferenceStaticType, DictionaryStaticType, RestrictedStaticType, TypeValue, and etc.

</details>

### Benchmark comparisons

Benchmarks using sample data from a mainnet snapshot, show:
 - Decoding time reduced by about 22% - 52%.
 - Memory alloc/op reduced by about 29% - 64%.
 - Memory allocs/op reduced by about 29% - 62%.
 
<details><summary>Composite values from mainnet (136-byte to 776155-byte)</summary>

```
name                                               old time/op    new time/op    delta
DecodeCBOR/comp_v2_96d7e06eaf4b3fcf_14850bytes-4      485µs ± 1%     284µs ± 1%  -41.42%  (p=0.000 n=9+9)
DecodeCBOR/comp_v3_99dc360eee32dcec_160bytes-4       11.4µs ± 1%     6.9µs ± 1%  -39.65%  (p=0.000 n=10+10)
DecodeCBOR/comp_v3_b52a33b7e56868f6_139bytes-4       8.05µs ± 1%    4.68µs ± 1%  -41.86%  (p=0.000 n=9+9)
DecodeCBOR/comp_v3_d99d7e6b4dad41e1_776155bytes-4    25.6ms ± 1%    20.0ms ± 1%  -21.89%  (p=0.000 n=9+9)

name                                               old alloc/op   new alloc/op   delta
DecodeCBOR/comp_v2_96d7e06eaf4b3fcf_14850bytes-4      196kB ± 0%     102kB ± 0%  -47.85%  (p=0.000 n=9+10)
DecodeCBOR/comp_v3_99dc360eee32dcec_160bytes-4       3.78kB ± 0%    2.10kB ± 0%  -44.49%  (p=0.000 n=10+10)
DecodeCBOR/comp_v3_b52a33b7e56868f6_139bytes-4       2.55kB ± 0%    1.30kB ± 0%  -49.22%  (p=0.000 n=10+10)
DecodeCBOR/comp_v3_d99d7e6b4dad41e1_776155bytes-4    10.1MB ± 0%     7.1MB ± 0%  -29.97%  (p=0.000 n=9+10)

name                                               old allocs/op  new allocs/op  delta
DecodeCBOR/comp_v2_96d7e06eaf4b3fcf_14850bytes-4      4.35k ± 0%     2.43k ± 0%  -44.23%  (p=0.000 n=10+10)
DecodeCBOR/comp_v3_99dc360eee32dcec_160bytes-4         85.0 ± 0%      47.0 ± 0%  -44.71%  (p=0.000 n=10+10)
DecodeCBOR/comp_v3_b52a33b7e56868f6_139bytes-4         52.0 ± 0%      28.0 ± 0%  -46.15%  (p=0.000 n=10+10)
DecodeCBOR/comp_v3_d99d7e6b4dad41e1_776155bytes-4      213k ± 0%      151k ± 0%  -29.31%  (p=0.000 n=9+10)
```

</details>

<details><summary>Link values from mainnet (167-byte, 192-byte)</summary>

```
name                                            old time/op    new time/op    delta
DecodeCBOR/link_v3_2392f05c3b72f235_167bytes-4    8.13µs ± 1%    3.91µs ± 0%  -51.92%  (p=0.000 n=10+10)
DecodeCBOR/link_v3_3a791fe1b8243e73_192bytes-4    8.13µs ± 1%    3.91µs ± 1%  -51.92%  (p=0.000 n=9+9)

name                                            old alloc/op   new alloc/op   delta
DecodeCBOR/link_v3_2392f05c3b72f235_167bytes-4    2.36kB ± 0%    0.84kB ± 0%  -64.41%  (p=0.000 n=10+10)
DecodeCBOR/link_v3_3a791fe1b8243e73_192bytes-4    2.36kB ± 0%    0.84kB ± 0%  -64.41%  (p=0.000 n=10+10)

name                                            old allocs/op  new allocs/op  delta
DecodeCBOR/link_v3_2392f05c3b72f235_167bytes-4      55.0 ± 0%      21.0 ± 0%  -61.82%  (p=0.000 n=10+10)
DecodeCBOR/link_v3_3a791fe1b8243e73_192bytes-4      55.0 ± 0%      21.0 ± 0%  -61.82%  (p=0.000 n=10+10)
```

</details>

### Caveats

DecodingCallback function used for debugging has changed:
 - The sequence of calls to the callback used to be top-down (parent then child).  Now they are bottom-up (child then parent).
 - The data passed to the callback used to be parsed by CBOR (to CBOR tags or native Go types).  Now the data is Value object.
 - TestDecodeCallback is updated to handle the new callback behavior.


______

For contributor use:

- [x] Targeted PR against `master` branch
- [x] Linked to Github issue with discussion and accepted design OR link to spec that describes this work
- [x] Code follows the [standards mentioned here](https://github.com/onflow/cadence/blob/master/CONTRIBUTING.md#styleguides)
- [x] Updated relevant documentation 
- [x] Re-reviewed `Files changed` in the Github PR explorer
- [x] Added appropriate labels